refactor(pylon): split 3518-line tests.rs into focused test modules
Split the monolithic tests.rs into 12 focused modules under tests/:
- auth (14 tests), error (15), error_envelope (26), health (12),
  helpers (shared fixtures), idempotency (7), message (13),
  middleware (22), nous (12), session (24), sse_events (18),
  streaming (8)

All 242 tests preserved. No individual file exceeds 500 lines.

Also fixes pre-existing clippy warnings in handlers/knowledge.rs:
- Prefix unused `state` param with underscore
- Gate `get_fact_relationships`/`get_similar_facts` behind
  `#[cfg(feature = "knowledge-store")]`

Closes #1421
