+++ crates/pylon/src/tests/health.rs
@@ -1,0 +1,245 @@
+use std::sync::Arc;
+
+use axum::body::Body;
+use axum::http::{Request, StatusCode};
+use tower::ServiceExt;
+
+use super::helpers::*;
+
+#[tokio::test]
+async fn health_no_auth_required() {
+    let (app, _dir) = app().await;
+    let resp = app
+        .oneshot(Request::get("/api/health").body(Body::empty()).unwrap())
+        .await
+        .unwrap();
+
+    assert_eq!(resp.status(), StatusCode::OK);
+    let body = body_json(resp).await;
+    assert_eq!(body["status"], "healthy");
+}
+
+#[tokio::test]
+async fn health_returns_200() {
+    let (app, _dir) = app().await;
+    let resp = app
+        .oneshot(Request::get("/api/health").body(Body::empty()).unwrap())
+        .await
+        .unwrap();
+
+    assert_eq!(resp.status(), StatusCode::OK);
+    let body = body_json(resp).await;
+    assert_eq!(body["status"], "healthy");
+    assert!(body["version"].is_string());
+    assert!(body["uptime_seconds"].is_number());
+    assert!(body["checks"].is_array());
+}
+
+#[tokio::test]
+async fn health_degraded_without_providers() {
+    let (app, _dir) = app_no_providers().await;
+    let resp = app
+        .oneshot(Request::get("/api/health").body(Body::empty()).unwrap())
+        .await
+        .unwrap();
+
+    let body = body_json(resp).await;
+    assert_eq!(body["status"], "degraded");
+}
+
+#[tokio::test]
+async fn health_checks_have_expected_shape() {
+    let (app, _dir) = app().await;
+    let resp = app
+        .oneshot(Request::get("/api/health").body(Body::empty()).unwrap())
+        .await
+        .unwrap();
+
+    assert_eq!(resp.status(), StatusCode::OK);
+    let body = body_json(resp).await;
+
+    let checks = body["checks"].as_array().expect("checks is array");
+    assert!(checks.len() >= 2, "expected at least 2 health checks");
+
+    for check in checks {
+        assert!(check["name"].is_string(), "each check has a name");
+        assert!(check["status"].is_string(), "each check has a status");
+    }
+
+    let names: Vec<&str> = checks.iter().filter_map(|c| c["name"].as_str()).collect();
+    assert!(
+        names.contains(&"session_store"),
+        "missing session_store check"
+    );
+    assert!(names.contains(&"providers"), "missing providers check");
+}
+
+// ── Metrics ─────────────────────────────────────────────────────────────────
+
+#[tokio::test]
+async fn metrics_returns_200_with_prometheus_content_type() {
+    let (app, _dir) = app().await;
+    let resp = app
+        .oneshot(Request::get("/metrics").body(Body::empty()).unwrap())
+        .await
+        .unwrap();
+
+    assert_eq!(resp.status(), StatusCode::OK);
+    let content_type = resp
+        .headers()
+        .get("content-type")
+        .unwrap()
+        .to_str()
+        .unwrap();
+    assert!(
+        content_type.contains("text/plain"),
+        "expected text/plain content type, got: {content_type}"
+    );
+}
+
+#[tokio::test]
+async fn metrics_no_auth_required() {
+    let (app, _dir) = app().await;
+    // No authorization header: should still succeed
+    let resp = app
+        .oneshot(Request::get("/metrics").body(Body::empty()).unwrap())
+        .await
+        .unwrap();
+
+    assert_eq!(resp.status(), StatusCode::OK);
+}
+
+#[tokio::test]
+async fn metrics_contains_aletheia_prefixed_families() {
+    let (app, _dir) = app().await;
+    let resp = app
+        .oneshot(Request::get("/metrics").body(Body::empty()).unwrap())
+        .await
+        .unwrap();
+
+    let body = body_string(resp).await;
+    assert!(
+        body.contains("aletheia_http_requests_total"),
+        "should contain HTTP request counter"
+    );
+    assert!(
+        body.contains("aletheia_uptime_seconds"),
+        "should contain uptime gauge"
+    );
+    assert!(
+        body.contains("# HELP"),
+        "should contain Prometheus HELP comments"
+    );
+    assert!(
+        body.contains("# TYPE"),
+        "should contain Prometheus TYPE comments"
+    );
+}
+
+#[tokio::test]
+async fn metrics_counters_increment_after_request() {
+    let (state, _dir) = test_state().await;
+    let router = build_router(Arc::clone(&state), &test_security_config());
+
+    // Make a health request first to increment the counter
+    let _ = router
+        .clone()
+        .oneshot(Request::get("/api/health").body(Body::empty()).unwrap())
+        .await
+        .unwrap();
+
+    // Then check /metrics for the counter
+    let resp = router
+        .clone()
+        .oneshot(Request::get("/metrics").body(Body::empty()).unwrap())
+        .await
+        .unwrap();
+
+    let body = body_string(resp).await;
+    assert!(
+        body.contains("/api/health"),
+        "should contain the health endpoint path in metrics"
+    );
+}
+
+// ── OpenAPI ─────────────────────────────────────────────────────────────────
+
+#[tokio::test]
+async fn openapi_spec_returns_valid_json() {
+    let (app, _dir) = app().await;
+    let resp = app
+        .oneshot(
+            Request::get("/api/docs/openapi.json")
+                .body(Body::empty())
+                .unwrap(),
+        )
+        .await
+        .unwrap();
+
+    assert_eq!(resp.status(), StatusCode::OK);
+    let body = body_json(resp).await;
+    let version = body["openapi"].as_str().unwrap();
+    assert!(
+        version.starts_with("3."),
+        "expected OpenAPI 3.x, got {version}"
+    );
+}
+
+#[tokio::test]
+async fn openapi_spec_has_all_paths() {
+    let (app, _dir) = app().await;
+    let resp = app
+        .oneshot(
+            Request::get("/api/docs/openapi.json")
+                .body(Body::empty())
+                .unwrap(),
+        )
+        .await
+        .unwrap();
+
+    let body = body_json(resp).await;
+    let paths = body["paths"].as_object().unwrap();
+    assert!(paths.contains_key("/api/health"));
+    assert!(paths.contains_key("/api/v1/sessions"));
+    assert!(paths.contains_key("/api/v1/sessions/{id}"));
+    assert!(paths.contains_key("/api/v1/sessions/{id}/messages"));
+    assert!(paths.contains_key("/api/v1/sessions/{id}/history"));
+    assert!(paths.contains_key("/api/v1/nous"));
+    assert!(paths.contains_key("/api/v1/nous/{id}"));
+    assert!(paths.contains_key("/api/v1/nous/{id}/tools"));
+}
+
+#[tokio::test]
+async fn openapi_docs_no_auth_required() {
+    let (app, _dir) = app().await;
+    let resp = app
+        .oneshot(
+            Request::get("/api/docs/openapi.json")
+                .body(Body::empty())
+                .unwrap(),
+        )
+        .await
+        .unwrap();
+
+    assert_eq!(resp.status(), StatusCode::OK);
+}
+
+#[tokio::test]
+async fn openapi_spec_has_schemas() {
+    let (app, _dir) = app().await;
+    let resp = app
+        .oneshot(
+            Request::get("/api/docs/openapi.json")
+                .body(Body::empty())
+                .unwrap(),
+        )
+        .await
+        .unwrap();
+
+    let body = body_json(resp).await;
+    let schemas = body["components"]["schemas"].as_object().unwrap();
+    assert!(schemas.contains_key("SessionResponse"));
+    assert!(schemas.contains_key("ErrorResponse"));
+    assert!(schemas.contains_key("HealthResponse"));
+    assert!(schemas.contains_key("NousStatus"));
+}
