+++ crates/pylon/src/tests/message.rs
@@ -1,0 +1,295 @@
+use std::sync::Arc;
+
+use axum::http::StatusCode;
+use tower::ServiceExt;
+
+use super::helpers::*;
+
+#[tokio::test]
+async fn send_message_returns_sse_content_type() {
+    let (state, _dir) = test_state().await;
+    let router = build_router(Arc::clone(&state), &test_security_config());
+    let created = create_test_session(&router).await;
+    let id = created["id"].as_str().unwrap();
+
+    let req = authed_request(
+        "POST",
+        &format!("/api/v1/sessions/{id}/messages"),
+        Some(serde_json::json!({ "content": "Hello!" })),
+    );
+
+    let resp = router.clone().oneshot(req).await.unwrap();
+    assert_eq!(resp.status(), StatusCode::OK);
+
+    let content_type = resp
+        .headers()
+        .get("content-type")
+        .unwrap()
+        .to_str()
+        .unwrap();
+    assert!(content_type.contains("text/event-stream"));
+}
+
+#[tokio::test]
+async fn send_message_stream_contains_events() {
+    let (state, _dir) = test_state().await;
+    let router = build_router(Arc::clone(&state), &test_security_config());
+    let created = create_test_session(&router).await;
+    let id = created["id"].as_str().unwrap();
+
+    let req = authed_request(
+        "POST",
+        &format!("/api/v1/sessions/{id}/messages"),
+        Some(serde_json::json!({ "content": "Hello!" })),
+    );
+
+    let resp = router.clone().oneshot(req).await.unwrap();
+    let body = body_string(resp).await;
+
+    assert!(
+        body.contains("event: text_delta"),
+        "should contain text_delta event"
+    );
+    assert!(
+        body.contains("Hello from mock!"),
+        "should contain mock response text"
+    );
+    assert!(
+        body.contains("event: message_complete"),
+        "should contain message_complete event"
+    );
+}
+
+#[tokio::test]
+async fn send_message_unknown_session_returns_404() {
+    let (app, _dir) = app().await;
+    let req = authed_request(
+        "POST",
+        "/api/v1/sessions/nonexistent/messages",
+        Some(serde_json::json!({ "content": "Hello!" })),
+    );
+
+    let resp = app.oneshot(req).await.unwrap();
+    assert_eq!(resp.status(), StatusCode::NOT_FOUND);
+}
+
+#[tokio::test]
+async fn send_empty_message_returns_400() {
+    let (router, _dir) = app().await;
+    let created = create_test_session(&router).await;
+    let id = created["id"].as_str().unwrap();
+
+    let req = authed_request(
+        "POST",
+        &format!("/api/v1/sessions/{id}/messages"),
+        Some(serde_json::json!({ "content": "" })),
+    );
+
+    let resp = router.clone().oneshot(req).await.unwrap();
+    assert_eq!(resp.status(), StatusCode::BAD_REQUEST);
+}
+
+#[tokio::test]
+async fn send_message_stores_in_history() {
+    let (state, _dir) = test_state().await;
+    let router = build_router(Arc::clone(&state), &test_security_config());
+    let created = create_test_session(&router).await;
+    let id = created["id"].as_str().unwrap();
+
+    let req = authed_request(
+        "POST",
+        &format!("/api/v1/sessions/{id}/messages"),
+        Some(serde_json::json!({ "content": "Hello!" })),
+    );
+    let resp = router.clone().oneshot(req).await.unwrap();
+    let _ = body_string(resp).await;
+
+    // Allow the spawned send_turn task to complete and store assistant message
+    tokio::time::sleep(std::time::Duration::from_millis(100)).await;
+
+    let resp = router
+        .clone()
+        .oneshot(authed_get(&format!("/api/v1/sessions/{id}/history")))
+        .await
+        .unwrap();
+
+    let body = body_json(resp).await;
+    let messages = body["messages"].as_array().unwrap();
+    assert!(messages.len() >= 2, "should have user + assistant messages");
+
+    assert_eq!(messages[0]["role"], "user");
+    assert_eq!(messages[0]["content"], "Hello!");
+}
+
+#[tokio::test]
+async fn empty_json_body_send_message_returns_400() {
+    let (router, _dir) = app().await;
+    let created = create_test_session(&router).await;
+    let id = created["id"].as_str().expect("session id");
+
+    let req = authed_request(
+        "POST",
+        &format!("/api/v1/sessions/{id}/messages"),
+        Some(serde_json::json!({})),
+    );
+
+    let resp = router.clone().oneshot(req).await.expect("response");
+    assert!(
+        resp.status() == StatusCode::BAD_REQUEST
+            || resp.status() == StatusCode::UNPROCESSABLE_ENTITY,
+        "expected 400 or 422, got {}",
+        resp.status()
+    );
+}
+
+#[tokio::test]
+async fn send_message_no_provider_returns_error() {
+    let (state, _dir) = test_state_with_provider(false).await;
+    let router = build_router(Arc::clone(&state), &test_security_config());
+    let created = create_test_session(&router).await;
+    let id = created["id"].as_str().unwrap();
+
+    let req = authed_request(
+        "POST",
+        &format!("/api/v1/sessions/{id}/messages"),
+        Some(serde_json::json!({ "content": "Hello!" })),
+    );
+
+    let resp = router.clone().oneshot(req).await.unwrap();
+    assert_eq!(resp.status(), StatusCode::INTERNAL_SERVER_ERROR);
+}
+
+#[tokio::test]
+async fn send_message_routes_through_actor() {
+    let (state, _dir) = test_state().await;
+    let router = build_router(Arc::clone(&state), &test_security_config());
+    let created = create_test_session(&router).await;
+    let id = created["id"].as_str().unwrap();
+
+    let req = authed_request(
+        "POST",
+        &format!("/api/v1/sessions/{id}/messages"),
+        Some(serde_json::json!({ "content": "Test routing" })),
+    );
+
+    let resp = router.clone().oneshot(req).await.unwrap();
+    assert_eq!(resp.status(), StatusCode::OK);
+    let body = body_string(resp).await;
+
+    assert!(body.contains("event: text_delta"), "should have text_delta");
+    assert!(
+        body.contains("Hello from mock!"),
+        "should contain mock response"
+    );
+    assert!(
+        body.contains("event: message_complete"),
+        "should have message_complete"
+    );
+    assert!(body.contains("end_turn"), "stop_reason should be end_turn");
+}
+
+// ── Stream turn ─────────────────────────────────────────────────────────────
+
+#[tokio::test]
+async fn stream_turn_returns_sse() {
+    let (state, _dir) = test_state().await;
+    let router = build_router(Arc::clone(&state), &test_security_config());
+
+    let req = authed_request(
+        "POST",
+        "/api/v1/sessions/stream",
+        Some(serde_json::json!({
+            "agentId": "syn",
+            "message": "Hello from TUI",
+            "sessionKey": "stream-test"
+        })),
+    );
+
+    let resp = router.oneshot(req).await.unwrap();
+    assert_eq!(resp.status(), StatusCode::OK);
+    let ct = resp
+        .headers()
+        .get("content-type")
+        .unwrap()
+        .to_str()
+        .unwrap();
+    assert!(ct.contains("text/event-stream"));
+}
+
+#[tokio::test]
+async fn stream_turn_contains_turn_start_event() {
+    let (state, _dir) = test_state().await;
+    let router = build_router(Arc::clone(&state), &test_security_config());
+
+    let req = authed_request(
+        "POST",
+        "/api/v1/sessions/stream",
+        Some(serde_json::json!({
+            "agentId": "syn",
+            "message": "Hello!",
+            "sessionKey": "stream-events"
+        })),
+    );
+
+    let resp = router.oneshot(req).await.unwrap();
+    let body = body_string(resp).await;
+    assert!(
+        body.contains("event: turn_start"),
+        "should contain turn_start event"
+    );
+}
+
+#[tokio::test]
+async fn stream_turn_empty_message_returns_400() {
+    let (app, _dir) = app().await;
+    let req = authed_request(
+        "POST",
+        "/api/v1/sessions/stream",
+        Some(serde_json::json!({
+            "agentId": "syn",
+            "message": "",
+            "sessionKey": "empty-msg"
+        })),
+    );
+
+    let resp = app.oneshot(req).await.unwrap();
+    assert_eq!(resp.status(), StatusCode::BAD_REQUEST);
+}
+
+#[tokio::test]
+async fn stream_turn_unknown_agent_returns_404() {
+    let (app, _dir) = app().await;
+    let req = authed_request(
+        "POST",
+        "/api/v1/sessions/stream",
+        Some(serde_json::json!({
+            "agentId": "nonexistent",
+            "message": "Hello!",
+            "sessionKey": "test"
+        })),
+    );
+
+    let resp = app.oneshot(req).await.unwrap();
+    assert_eq!(resp.status(), StatusCode::NOT_FOUND);
+}
+
+// ── Events endpoint ─────────────────────────────────────────────────────────
+
+#[tokio::test]
+async fn events_endpoint_returns_200_sse() {
+    // /api/v1/events must return 200 with SSE content-type (#1248).
+    let (app, _dir) = app().await;
+    let resp = app.oneshot(authed_get("/api/v1/events")).await.unwrap();
+
+    assert_eq!(resp.status(), StatusCode::OK);
+    let ct = resp
+        .headers()
+        .get("content-type")
+        .unwrap()
+        .to_str()
+        .unwrap();
+    assert!(
+        ct.contains("text/event-stream"),
+        "expected text/event-stream content-type, got: {ct}"
+    );
+}
