+++ crates/pylon/src/tests/nous.rs
@@ -1,0 +1,178 @@
+use std::sync::Arc;
+
+use axum::http::StatusCode;
+use tower::ServiceExt;
+
+use super::helpers::*;
+
+#[tokio::test]
+async fn list_nous_returns_agents() {
+    let (app, _dir) = app().await;
+    let resp = app.oneshot(authed_get("/api/v1/nous")).await.unwrap();
+
+    assert_eq!(resp.status(), StatusCode::OK);
+    let body = body_json(resp).await;
+    let agents = body["nous"].as_array().unwrap();
+    assert_eq!(agents.len(), 1);
+    assert_eq!(agents[0]["id"], "syn");
+}
+
+#[tokio::test]
+async fn get_nous_status() {
+    let (app, _dir) = app().await;
+    let resp = app.oneshot(authed_get("/api/v1/nous/syn")).await.unwrap();
+
+    assert_eq!(resp.status(), StatusCode::OK);
+    let body = body_json(resp).await;
+    assert_eq!(body["id"], "syn");
+    assert!(body["context_window"].is_number());
+    assert!(body["max_output_tokens"].is_number());
+}
+
+#[tokio::test]
+async fn get_unknown_nous_returns_404() {
+    let (app, _dir) = app().await;
+    let resp = app
+        .oneshot(authed_get("/api/v1/nous/nonexistent"))
+        .await
+        .unwrap();
+
+    assert_eq!(resp.status(), StatusCode::NOT_FOUND);
+    let body = body_json(resp).await;
+    assert_eq!(body["error"]["code"], "nous_not_found");
+}
+
+#[tokio::test]
+async fn get_nous_tools() {
+    let (app, _dir) = app().await;
+    let resp = app
+        .oneshot(authed_get("/api/v1/nous/syn/tools"))
+        .await
+        .unwrap();
+
+    assert_eq!(resp.status(), StatusCode::OK);
+    let body = body_json(resp).await;
+    assert!(body["tools"].is_array());
+}
+
+#[tokio::test]
+async fn nous_list_from_manager() {
+    let (state, _dir) = test_state().await;
+    let router = build_router(Arc::clone(&state), &test_security_config());
+
+    let resp = router.oneshot(authed_get("/api/v1/nous")).await.unwrap();
+
+    assert_eq!(resp.status(), StatusCode::OK);
+    let body = body_json(resp).await;
+    let agents = body["nous"].as_array().unwrap();
+    assert_eq!(agents.len(), 1);
+    assert_eq!(agents[0]["id"], "syn");
+    assert_eq!(agents[0]["model"], "mock-model");
+    assert_eq!(agents[0]["status"], "active");
+}
+
+#[tokio::test]
+async fn nous_tools_unknown_returns_404() {
+    let (app, _dir) = app().await;
+    let resp = app
+        .oneshot(authed_get("/api/v1/nous/nonexistent/tools"))
+        .await
+        .unwrap();
+
+    assert_eq!(resp.status(), StatusCode::NOT_FOUND);
+    let body = body_json(resp).await;
+    assert_eq!(body["error"]["code"], "nous_not_found");
+}
+
+#[tokio::test]
+async fn nous_status_response_has_all_fields() {
+    let (app, _dir) = app().await;
+    let resp = app.oneshot(authed_get("/api/v1/nous/syn")).await.unwrap();
+
+    let body = body_json(resp).await;
+    assert!(body["id"].is_string());
+    assert!(body["model"].is_string());
+    assert!(body["context_window"].is_number());
+    assert!(body["max_output_tokens"].is_number());
+    assert!(body["thinking_enabled"].is_boolean());
+    assert!(body["thinking_budget"].is_number());
+    assert!(body["max_tool_iterations"].is_number());
+    assert!(body["status"].is_string());
+}
+
+// ── Config endpoints ────────────────────────────────────────────────────────
+
+#[tokio::test]
+async fn config_get_returns_redacted_config() {
+    let (app, _dir) = app().await;
+    let resp = app.oneshot(authed_get("/api/v1/config")).await.unwrap();
+
+    assert_eq!(resp.status(), StatusCode::OK);
+    let body = body_json(resp).await;
+    assert!(body.is_object());
+}
+
+#[tokio::test]
+async fn config_get_section_valid() {
+    let (app, _dir) = app().await;
+    let resp = app
+        .oneshot(authed_get("/api/v1/config/gateway"))
+        .await
+        .unwrap();
+
+    assert_eq!(resp.status(), StatusCode::OK);
+}
+
+#[tokio::test]
+async fn config_get_section_invalid_returns_404() {
+    let (app, _dir) = app().await;
+    let resp = app
+        .oneshot(authed_get("/api/v1/config/nonexistent_section"))
+        .await
+        .unwrap();
+
+    assert_eq!(resp.status(), StatusCode::NOT_FOUND);
+}
+
+#[tokio::test]
+async fn config_update_invalid_section_returns_404() {
+    let (app, _dir) = app().await;
+    let req = authed_request(
+        "PUT",
+        "/api/v1/config/nonexistent_section",
+        Some(serde_json::json!({"key": "value"})),
+    );
+
+    let resp = app.oneshot(req).await.unwrap();
+    assert_eq!(resp.status(), StatusCode::NOT_FOUND);
+}
+
+#[tokio::test]
+async fn gateway_config_signing_key_is_redacted() {
+    let (state, _dir) = test_state().await;
+
+    // Inject a signing key so there is a secret value to redact.
+    {
+        let mut config = state.config.write().await;
+        config.gateway.auth.signing_key = Some("super-secret-signing-key".to_owned());
+    }
+
+    let router = build_router(Arc::clone(&state), &test_security_config());
+    let resp = router
+        .oneshot(authed_get("/api/v1/config/gateway"))
+        .await
+        .unwrap();
+
+    assert_eq!(resp.status(), StatusCode::OK);
+    let body = body_json(resp).await;
+
+    // The raw secret must not appear anywhere in the response.
+    assert!(
+        !body.to_string().contains("super-secret-signing-key"),
+        "signing key must not appear in API response"
+    );
+    // The field must be replaced with the redaction placeholder.
+    assert_eq!(body["auth"]["signingKey"], "***");
+    // Non-secret fields must still be present and correct.
+    assert_eq!(body["port"], 18789);
+}
